nationality test 2

+++ app/Http/Controllers/nationalityController.php
@@ -17,6 +17,7 @@ class nationalityController extends Controller
 //        $this->middleware(['permission:nationalities_delete'])->only('destroy');
 
     }
+
     /**
      * Display a listing of the resource.
      *
@@ -41,57 +42,57 @@ class nationalityController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
         $validator = validator($request->all(), [
-            "name" =>  "required",
+            "name" => "required",
         ]);
 
-        if($validator->fails()) {
-            return responseJson(0, $validator->errors()->getMessages(),""  );
+        if ($validator->fails()) {
+            return responseJson(0, $validator->errors()->getMessages(), "");
         }
         try {
             $nationality = Nationality::create($request->all());
-            if($nationality){
-                notify()->success("تم حفظ البيانات بنجاح","","bottomLeft",);
+            if ($nationality) {
+                notify()->success("تم حفظ البيانات بنجاح", "", "bottomLeft",);
 
-                return responseJson(1, __('ok'),$nationality  );
-            }else{
+                return responseJson(1, __('ok'), $nationality);
+            } else {
             }
         } catch (\Exception $ex) {
-            return responseJson(0, "",$ex->getMessage()  );
+            return responseJson(0, "", $ex->getMessage());
         }
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function show($id)
     {
         $nationality = Nationality::find($id);
         if (!$nationality) {
-            return responseJson(0, __('data not found'),''  );
+            return responseJson(0, __('data not found'), '');
         }
-        return responseJson(1, "ok",$nationality  );
+        return responseJson(1, "ok", $nationality);
     }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
     {
         $nationality = Nationality::find($id);
         if (!$nationality) {
-            notify()->warning("هذه البيانات غير موجودة !","","bottomLeft",);
+            notify()->warning("هذه البيانات غير موجودة !", "", "bottomLeft",);
             return redirect()->route('nationalities.index');
         }
         return view('main_settings.nationalities.edit', compact('nationality'));
@@ -101,8 +102,8 @@ class nationalityController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param \Illuminate\Http\Request $request
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function update(NationalityRequest $request, $id)
@@ -110,22 +111,22 @@ class nationalityController extends Controller
         try {
             $nationality = Nationality::find($id);
             if (!$nationality) {
-                return responseJson(0, __('data not found'),''  );
+                return responseJson(0, __('data not found'), '');
 
             } else {
                 $nationality->update($request->all());
-                return responseJson(1, __('data updated successfully'), $nationality );
+                return responseJson(1, __('data updated successfully'), $nationality);
 
             }
         } catch (\Exception $ex) {
-            return responseJson(0, "",$ex->getMessage()  );
+            return responseJson(0, "", $ex->getMessage());
         }
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
@@ -133,12 +134,18 @@ class nationalityController extends Controller
         try {
             $nationality = Nationality::find($id);
             if (!$nationality) {
-                return responseJson(0, __('data not found'),''  );
+                return responseJson(0, __('data not found'), '');
+            }
+
+            if ($nationality->applications->count() > 0 || $nationality->students->count()  > 0 ) {
+                return responseJson(0, __('this item can not be deleted'), $nationality->fresh());
+
             }
             $nationality->delete();
-            return responseJson(1, __('deleted successfully'),''  );
+            return responseJson(1, __('deleted successfully'), '');
         } catch (\Exception $ex) {
-            return responseJson(0, "",$ex->getMessage()  );
+            return responseJson(0, "", $ex->getMessage());
         }
     }
+
 }
